test: ServiceTestBase에서 Service 호출 로직 제거
